Fix bug with circle imports of tools from waggon.utils
