+++ waggon/utils/__init__.py
@@ -1,6 +1,4 @@
-from .utils import display
 from .utils import plot_results
-from .utils import plot_func_2d
 
 from .utils import set_all_seed
 from .utils import fixed_random_seed
@@ -10,9 +8,7 @@ from .utils import fixed_all_seed
 
 
 __all__ = [
-    'display',
     'plot_results',
-    'plot_func_2d',
     'set_all_seed',
     'fixed_random_seed',
     'fixed_numpy_seed',
